+++ Client/src/main/java/module-info.java
@@ -4,7 +4,7 @@ module cs3321.client {
     requires java.net.http;
     requires javafx.graphics;
     requires Game;
-
+    requires com.google.gson;
 
     opens cs3321.client to javafx.fxml;
     exports cs3321.client;